Modify: Basemodel modularize
