+++ taipei-day-trip/app/app.py
@@ -2,7 +2,6 @@
 from fastapi.staticfiles import StaticFiles
 from fastapi.responses import FileResponse, JSONResponse
 from fastapi import *
-from fastapi.encoders import jsonable_encoder
 from fastapi.exceptions import RequestValidationError
 from starlette.exceptions import HTTPException as StarletteHTTPException
 
@@ -55,7 +54,8 @@ async def thankyou(request: Request):
 
 # Error handler
 
-
+# fastAPI - When a request contains invalid data, FastAPI internally raises a RequestValidationError
+# RequestValidationError is a sub-class of Pydantic's ValidationError
 @app.exception_handler(RequestValidationError)
 async def validation_exception_handler(request: Request, exc: RequestValidationError):
     return JSONResponse(
@@ -66,7 +66,10 @@ async def validation_exception_handler(request: Request, exc: RequestValidationE
         }
     )
 
-
+# And FastAPI's HTTPException error class inherits from Starlette's HTTPException error class.
+# The only difference is that FastAPI's HTTPException accepts any JSON-able data for the detail field, while Starlette's HTTPException only accepts strings for it.
+# But when you register an exception handler, you should register it for Starlette's HTTPException.
+# This way, if any part of Starlette's internal code, or a Starlette extension or plug-in, raises a Starlette HTTPException, your handler will be able to catch and handle it
 @app.exception_handler(StarletteHTTPException)
 async def http_exception_handler(request, exc):
     return JSONResponse(
