+++ taipei-day-trip/app/routers/user.py
@@ -1,11 +1,10 @@
 from fastapi import APIRouter, Request, Header
 from fastapi.responses import JSONResponse
-from pydantic import BaseModel, EmailStr, Field, field_validator
-import re
 import datetime
 import jwt
 from starlette import status
 from app.function.function import *
+from app.function.basemodel import *
 from dotenv import load_dotenv
 import os
 import sys
@@ -19,52 +18,6 @@ router = APIRouter(
 )
 
 # User registration
-
-## Regex pattern
-
-pattern_name = r'^[a-zA-Z0-9]{3,30}$'
-pattern_password = r'^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)[a-zA-Z\d]{8,20}$'
-
-## Class
-
-class UserSignUPInput(BaseModel):
-    name: str = Field(..., min_length=3, max_length=30,
-                      pattern=pattern_name)
-    email: EmailStr
-    password: str = Field(..., min_length=8, max_length=20,)
-
-    @field_validator('password')
-    def password_pattern(cls, v):
-        if re.match(pattern_password, v) is None:
-            raise ValueError(
-                'password should contain at least 1 uppercase, 1 lowercase and 1 number.')
-        return v
-
-
-class UserSignInInput(BaseModel):
-    email: EmailStr = Field(..., max_length=50)
-    password: str = Field(..., min_length=8, max_length=20,)
-
-    @field_validator('password')
-    def password_pattern(cls, v):
-        if re.match(pattern_password, v) is None:
-            raise ValueError(
-                'password should contain at least 1 uppercase, 1 lowercase and 1 number.')
-        return v
-
-
-class User(BaseModel):
-    id: int
-    name: str = Field(..., min_length=3, max_length=30,
-                      pattern=pattern_name)
-    email: EmailStr
-
-
-class ReturnUser(BaseModel):
-    data: User | None
-
-## path
-
 
 @router.post("", response_class=JSONResponse, summary="註冊一個新的會員")
 async def sign_up(user: UserSignUPInput):
